pyproject.toml
[tool.poetry]
name = "blog_app"
version = "0.1.0"
description = ""
authors = ["Te-jé Rodgers <tjd.rodgers@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = "^0.49.1"
auth0-python = "^3.14.0"
typed-settings = "^0.9.2"
PyJWT = "^1"
SQLAlchemy = "1.4.0b3"
aiomysql = "^0.0.21"
cryptography = "^3.4.6"
bson = "^0.5.10"
aiostream = "^0.4.2"
starlette = {version = "^0.14.2", optional = true}
pytest = {version = "^6.2.2", optional = true}
pytest-cov = {version = "^2.11.1", optional = true}
pytest-mock = {version = "^3.5.1", optional = true}
pytest-asyncio = {version = "^0.14.0", optional = true}
pytest-factoryboy = {version = "^2.1.0", optional = true}
requests-mock = {version = "^1.8.0", optional = true}
uvicorn = {extras = ["standard"], version = "^0.13.4", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
black = "^20.8b1"
mypy = "^0.812"
typer = "^0.3.2"
pytest-cov = "^2.11.1"
pytest-mock = "^3.5.1"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
pytest-asyncio = "^0.14.0"
requests-mock = "^1.8.0"
sqlalchemy-stubs = "^0.4"
pytest-factoryboy = "^2.1.0"

[tool.poetry.extras]
test = ["starlette", "pytest", "pytest-cov", "pytest-mock", "pytest-asyncio", "pytest-factoryboy", "requests-mock"]
server = ["uvicorn"]

[tool.poetry.scripts]
devtools = "cli.devtools:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
